Rename sierra_bibs_to_dynamo to sierra_reader
