+++ sierra_adapter/sierra_reader/Dockerfile
