+++ sierra_adapter/sierra_reader/src/main/resources/logback.xml
