+++ sierra_adapter/sierra_reader/src/main/scala/uk/ac/wellcome/platform/sierra_reader/Server.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_bibs_to_sns
+package uk.ac.wellcome.platform.sierra_reader
 
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.HttpServer
@@ -10,13 +10,13 @@ import com.twitter.finatra.http.filters.{
 import com.twitter.finatra.http.routing.HttpRouter
 import uk.ac.wellcome.finatra.modules._
 import uk.ac.wellcome.finatra.controllers.ManagementController
-import uk.ac.wellcome.platform.sierra_bibs_to_sns.modules.SierraBibsToSnsModule
+import uk.ac.wellcome.platform.sierra_reader.modules.SierraBibsToSnsModule
 
 object ServerMain extends Server
 
 class Server extends HttpServer {
   override val name =
-    "uk.ac.wellcome.platform.sierra_bibs_to_sns SierraBibsToSNS"
+    "uk.ac.wellcome.platform.sierra_reader SierraReader"
   override val modules = Seq(
     SierraBibsToSnsModule,
     AmazonCloudWatchModule,
