+++ sierra_adapter/sierra_reader/src/main/scala/uk/ac/wellcome/platform/sierra_reader/services/SierraBibsToSnsWorkerService.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_bibs_to_sns.services
+package uk.ac.wellcome.platform.sierra_reader.services
 
 import akka.Done
 import akka.actor.ActorSystem
@@ -7,7 +7,7 @@ import com.google.inject.Inject
 import com.twitter.inject.annotations.Flag
 import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.aws.SQSMessage
-import uk.ac.wellcome.platform.sierra_bibs_to_sns.sink.SierraBibsToSnsSink
+import uk.ac.wellcome.platform.sierra_reader.sink.SierraBibsToSnsSink
 import uk.ac.wellcome.sierra.{SierraSource, ThrottleRate}
 import uk.ac.wellcome.sns.SNSWriter
 import uk.ac.wellcome.sqs.{SQSReader, SQSWorker}
