+++ sierra_adapter/sierra_reader/src/main/scala/uk/ac/wellcome/platform/sierra_reader/sink/SierraBibsToSnsSink.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_bibs_to_sns.sink
+package uk.ac.wellcome.platform.sierra_reader.sink
 
 import java.time.format.DateTimeFormatter
 import java.time.{LocalDate, ZoneOffset}
