+++ sierra_adapter/sierra_reader/src/test/resources/logback.xml
