+++ sierra_adapter/sierra_reader/src/test/resources/wiremock/mappings/token-8be78772-0670-4bf7-bc54-dd9997eb6c74.json
