+++ sierra_adapter/sierra_reader/src/test/scala/uk/ac/wellcome/platform/sierra_reader/sink/SierraBibsToSnsSinkTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_bibs_to_sns.sink
+package uk.ac.wellcome.platform.sierra_reader.sink
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
