+++ sierra_adapter/sierra_reader/src/universal/conf/application.ini.template
